+++ packages/shopify-app-remix/docs/staticPages/examples/index/splat-route.example.ts
@@ -1,8 +1,8 @@
-import {LoaderArgs} from '@remix-run/node';
+import {LoaderFunctionArgs} from '@remix-run/node';
 
 import shopify from '~/shopify.server';
 
-export async function loader({request}: LoaderArgs) {
+export async function loader({request}: LoaderFunctionArgs) {
   await shopify.authenticate.admin(request);
 
   // App logic goes here