+++ packages/shopify-app-remix/src/server/authenticate/admin/__tests__/session-token-header-path.test.ts
@@ -113,11 +113,9 @@ describe('authorize.session token header path', () => {
 
       it('returns context when session exists for non-embedded apps', async () => {
         // GIVEN
-        const shopify = shopifyApp({
-          ...testConfig(),
-          isEmbeddedApp: false,
-          useOnlineTokens: isOnline,
-        });
+        const shopify = shopifyApp(
+          testConfig({isEmbeddedApp: false, useOnlineTokens: isOnline}),
+        );
 
         let testSession: Session;
         testSession = await setUpValidSession(shopify.sessionStorage);