+++ packages/shopify-app-remix/src/server/authenticate/admin/billing/types.ts
@@ -25,6 +25,14 @@ export interface RequestBillingOptions<Config extends AppConfigArg>
    * The plan to request. Must be one of the values defined in the `billing` config option.
    */
   plan: keyof Config['billing'];
+  /**
+   * Whether to use the test mode. This prevents the credit card from being charged. Test shops and demo shops cannot be charged.
+   */
+  isTest?: boolean;
+  /**
+   * The URL to return to after the merchant approves the payment.
+   */
+  returnUrl?: string;
 }
 
 export interface CancelBillingOptions {
@@ -38,6 +46,9 @@ export interface CancelBillingOptions {
    * {@link https://shopify.dev/docs/apps/billing/subscriptions/cancel-recurring-charges}
    */
   prorate?: boolean;
+  /*
+   * Whether to use the test mode. This prevents the credit card from being charged. Test shops and demo shops cannot be charged.
+   */
   isTest?: boolean;
 }
 
@@ -49,13 +60,13 @@ export interface BillingContext<Config extends AppConfigArg> {
    *
    * @example
    * <caption>Requesting billing right away.</caption>
-   * <description>Call `billing.request` in the `onFailure` callback to immediately request payment.</description>
+   * <description>Call `billing.request` in the `onFailure` callback to immediately redirect to the Shopify page to request payment.</description>
    * ```ts
    * // /app/routes/**\/*.ts
-   * import { LoaderArgs } from "@remix-run/node";
+   * import { LoaderFunctionArgs } from "@remix-run/node";
    * import { authenticate, MONTHLY_PLAN } from "../shopify.server";
    *
-   * export const loader = async ({ request }: LoaderArgs) => {
+   * export const loader = async ({ request }: LoaderFunctionArgs) => {
    *   const { billing } = await authenticate.admin(request);
    *   await billing.require({
    *     plans: [MONTHLY_PLAN],
@@ -94,13 +105,13 @@ export interface BillingContext<Config extends AppConfigArg> {
    *
    * @example
    * <caption>Using a plan selection page.</caption>
-   * <description>Redirect to a different page in the `onFailure` callback, where the merchant can select a billing plan.</description>
+   * <description> When the app has multiple plans, create a page in your App that allows the merchant to select a plan. If a merchant does not have the required plan you can redirect them to page in your app to select one.</description>
    * ```ts
    * // /app/routes/**\/*.ts
-   * import { LoaderArgs, redirect } from "@remix-run/node";
+   * import { LoaderFunctionArgs, redirect } from "@remix-run/node";
    * import { authenticate, MONTHLY_PLAN, ANNUAL_PLAN } from "../shopify.server";
    *
-   * export const loader = async ({ request }: LoaderArgs) => {
+   * export const loader = async ({ request }: LoaderFunctionArgs) => {
    *   const { billing } = await authenticate.admin(request);
    *   const billingCheck = await billing.require({
    *     plans: [MONTHLY_PLAN, ANNUAL_PLAN],
@@ -154,17 +165,17 @@ export interface BillingContext<Config extends AppConfigArg> {
    * <description>Change where the merchant is returned to after approving the purchase using the `returnUrl` option.</description>
    * ```ts
    * // /app/routes/**\/*.ts
-   * import { LoaderArgs } from "@remix-run/node";
+   * import { LoaderFunctionArgs } from "@remix-run/node";
    * import { authenticate, MONTHLY_PLAN } from "../shopify.server";
    *
-   * export const loader = async ({ request }: LoaderArgs) => {
+   * export const loader = async ({ request }: LoaderFunctionArgs) => {
    *   const { billing } = await authenticate.admin(request);
    *   await billing.require({
    *     plans: [MONTHLY_PLAN],
    *     onFailure: async () => billing.request({
    *       plan: MONTHLY_PLAN,
    *       isTest: true,
-   *       returnUrl: '/billing-complete',
+   *       returnUrl: 'https://admin.shopify.com/store/my-store/apps/my-app/billing-page',
    *     }),
    *   });
    *
@@ -209,10 +220,10 @@ export interface BillingContext<Config extends AppConfigArg> {
    * <description>Use the `billing.cancel` function to cancel an active subscription with the id returned from `billing.require`.</description>
    * ```ts
    * // /app/routes/cancel-subscription.ts
-   * import { LoaderArgs } from "@remix-run/node";
+   * import { LoaderFunctionArgs } from "@remix-run/node";
    * import { authenticate, MONTHLY_PLAN } from "../shopify.server";
    *
-   * export const loader = async ({ request }: LoaderArgs) => {
+   * export const loader = async ({ request }: LoaderFunctionArgs) => {
    *   const { billing } = await authenticate.admin(request);
    *   const billingCheck = await billing.require({
    *     plans: [MONTHLY_PLAN],