+++ packages/shopify-app-remix/src/server/clients/admin/types.ts
@@ -29,7 +29,19 @@ export interface AdminApiContext<
    *
    * @example
    * <caption>Using REST resources.</caption>
-   * <description>Getting the number of orders in a store using REST resources.</description>
+   * <description>Getting the number of orders in a store using REST resources. Visit the [Admin REST API references](/docs/api/admin-rest) for examples on using each resource. </description>
+   *
+   * ```ts
+   * // /app/routes/**\/*.ts
+   * import { LoaderFunctionArgs, json } from "@remix-run/node";
+   * import { authenticate } from "../shopify.server";
+   *
+   * export const loader = async ({ request }: LoaderFunctionArgs) => {
+   *   const { admin, session } = await authenticate.admin(request);
+   *   return json(admin.rest.resources.Order.count({ session }));
+   * };
+   * ```
+   *
    *
    * ```ts
    * // /app/shopify.server.ts
@@ -44,20 +56,23 @@ export interface AdminApiContext<
    * export const authenticate = shopify.authenticate;
    * ```
    *
-   * ```ts
-   * // /app/routes/**\/*.ts
-   * import { LoaderArgs, json } from "@remix-run/node";
-   * import { authenticate } from "../shopify.server";
-   *
-   * export const loader = async ({ request }: LoaderArgs) => {
-   *   const { admin, session } = await authenticate.admin(request);
-   *   return json(admin.rest.resources.Order.count({ session }));
-   * };
-   * ```
    *
    * @example
    * <caption>Performing a GET request to the REST API.</caption>
-   * <description>Use `admin.rest.<method>` to make custom requests to the API.</description>
+   * <description>Use `admin.rest.get` to make custom requests to make a request to to the `customer/count` endpoint</description>
+   *
+   * ```ts
+   * // /app/routes/**\/*.ts
+   * import { LoaderFunctionArgs, json } from "@remix-run/node";
+   * import { authenticate } from "../shopify.server";
+   *
+   * export const loader = async ({ request }: LoaderFunctionArgs) => {
+   *   const { admin, session } = await authenticate.admin(request);
+   *   const response = await admin.rest.get({ path: "/customers/count.json" });
+   *   const customers = await response.json();
+   *   return json({ customers });
+   * };
+   * ```
    *
    * ```ts
    * // /app/shopify.server.ts
@@ -71,18 +86,44 @@ export interface AdminApiContext<
    * export default shopify;
    * export const authenticate = shopify.authenticate;
    * ```
-   *
+   * @example
+   * <caption>Performing a POST request to the REST API.</caption>
+   * <description>Use `admin.rest.post` to make custom requests to make a request to to the `customers.json` endpoint to send a welcome email</description>
    * ```ts
    * // /app/routes/**\/*.ts
-   * import { LoaderArgs, json } from "@remix-run/node";
+   * import { LoaderFunctionArgs, json } from "@remix-run/node";
    * import { authenticate } from "../shopify.server";
    *
-   * export const loader = async ({ request }: LoaderArgs) => {
+   * export const loader = async ({ request }: LoaderFunctionArgs) => {
    *   const { admin, session } = await authenticate.admin(request);
-   *   const response = await admin.rest.get({ path: "/customers/count.json" });
-   *   const customers = await response.json();
-   *   return json({ customers });
+   *   const response = admin.rest.post({
+   *     path: "customers/7392136888625/send_invite.json",
+   *     body: {
+   *       customer_invite: {
+   *         to: "new_test_email@shopify.com",
+   *         from: "j.limited@example.com",
+   *         bcc: ["j.limited@example.com"],
+   *         subject: "Welcome to my new shop",
+   *         custom_message: "My awesome new store",
+   *       },
+   *     },
+   * });
+   *   const customerInvite = await response.json();
+   *   return json({ customerInvite });
    * };
+   * ```
+   *
+   * ```ts
+   * // /app/shopify.server.ts
+   * import { shopifyApp } from "@shopify/shopify-app-remix/server";
+   * import { restResources } from "@shopify/shopify-api/rest/admin/2023-04";
+   *
+   * const shopify = shopifyApp({
+   *   restResources,
+   *   // ...etc
+   * });
+   * export default shopify;
+   * export const authenticate = shopify.authenticate;
    * ```
    */
   rest: RestClientWithResources<Resources>;
@@ -91,16 +132,16 @@ export interface AdminApiContext<
    * Methods for interacting with the Shopify Admin GraphQL API
    *
    * {@link https://shopify.dev/docs/api/admin-graphql}
-   * {@link https://github.com/Shopify/shopify-api-js/blob/main/docs/reference/clients/Graphql.md}
+   * {@link https://github.com/Shopify/shopify-api-js/blob/main/packages/shopify-api/docs/reference/clients/Graphql.md}
    *
    * @example
    * <caption>Querying the GraphQL API.</caption>
    * <description>Use `admin.graphql` to make query / mutation requests.</description>
    * ```ts
-   * import { ActionArgs } from "@remix-run/node";
+   * import { ActionFunctionArgs } from "@remix-run/node";
    * import { authenticate } from "../shopify.server";
    *
-   * export async function action({ request }: ActionArgs) {
+   * export async function action({ request }: ActionFunctionArgs) {
    *   const { admin } = await authenticate.admin(request);
    *
    *   const response = await admin.graphql(