+++ packages/shopify-app-session-storage-memory/loom.config.ts
@@ -9,7 +9,7 @@ export default createPackage((pkg) => {
       // Use browser targets (e.g. `'defaults'`) if your package targets the browser,
       // node targets (e.g. `'node 12.22'`) if your package targets node
       // or both (e.g.`'defaults, node 12.22'`) if your package targets both
-      targets: 'node 14',
+      targets: 'node 16',
       // Optional. Defaults to false. Defines if commonjs outputs should be generated.
       commonjs: true,
       // Optional. Defaults to false. Defines if esmodules outputs should be generated.