+++ packages/shopify-app-session-storage-mongodb/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @shopify/shopify-app-session-storage-mongodb
 
+## 2.0.1
+
+### Patch Changes
+
+- 6d12840: Updating dependencies on @shopify/shopify-api
+- Updated dependencies [6d12840]
+  - @shopify/shopify-app-session-storage@2.0.1
+
+## 2.0.0
+
+### Major Changes
+
+- f837060: **Removed support for Node 14**
+
+  Node 14 has reached its [EOL](https://endoflife.date/nodejs), and dependencies to this package no longer work on Node 14.
+  Because of that, we can no longer support that version.
+
+  If your app is running on Node 14, you'll need to update to a more recent version before upgrading this package.
+
+  This upgrade does not require any code changes.
+
+### Patch Changes
+
+- a69d6fc: Updating dependency on @shopify/shopify-api to v.8.0.1
+- Updated dependencies [f837060]
+- Updated dependencies [a69d6fc]
+  - @shopify/shopify-app-session-storage@2.0.0
+
 ## 1.0.15
 
 ### Patch Changes
@@ -69,7 +97,7 @@
 - 7afa957: Bumps [mongodb](https://github.com/mongodb/node-mongodb-native) from 5.2.0 to 5.3.0. See mongodb's [changelog](https://github.com/mongodb/node-mongodb-native/blob/main/HISTORY.md) for more details.
 - 9166b94: Bumps [mongodb](https://github.com/mongodb/node-mongodb-native) from 5.2.0 to 5.4.0. See mongodb's [changelog](https://github.com/mongodb/node-mongodb-native/blob/main/HISTORY.md) for more details.
 - e1d4f4f: Add @shopify/shopify-api as a peerDependencies entry for each session-storage package, to avoid API library conflicts (e.g., scopesArray.map error). Should help avoid issues like #93
-- 1d007e8: Bumps [@shopify/shopify-api](https://github.com/Shopify/shopify-api-js) from 7.0.0 to 7.1.0. See `@shopify/shopify-api`'s [changelog](https://github.com/Shopify/shopify-api-js/blob/main/CHANGELOG.md) for more details.
+- 1d007e8: Bumps [@shopify/shopify-api](https://github.com/Shopify/shopify-api-js) from 7.0.0 to 7.1.0. See `@shopify/shopify-api`'s [changelog](https://github.com/Shopify/shopify-api-js/blob/main/packages/shopify-api/CHANGELOG.md) for more details.
 - Updated dependencies [e1d4f4f]
 - Updated dependencies [1d007e8]
   - @shopify/shopify-app-session-storage@1.1.3
@@ -78,7 +106,7 @@
 
 ### Patch Changes
 
-- e4f3415: Bump @shopify/shopify-api from 6.2.0 to 7.0.0. See [changelog](https://github.com/Shopify/shopify-api-js/blob/main/CHANGELOG.md) for details.
+- e4f3415: Bump @shopify/shopify-api from 6.2.0 to 7.0.0. See [changelog](https://github.com/Shopify/shopify-api-js/blob/main/packages/shopify-api/CHANGELOG.md) for details.
 - 7bdb422: Bumps mongodb from 5.1.0 to 5.2.0. See mongodb [changelog](https://github.com/mongodb/node-mongodb-native/blob/main/HISTORY.md) for details.
 - e571c35: Bump mongodb from 5.0.1 to 5.1.0. See mongodb [changelog](https://github.com/mongodb/node-mongodb-native/blob/main/HISTORY.md) for details.
 - Updated dependencies [e4f3415]